+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Travis CI
 - Support IE9+ (previously IE10+) thanks in part to [html5-history-api](https://github.com/devote/HTML5-History-API)
 - Now will work in web workers and pretty much anywhere that runs js.
+- Server#Navigate now has the same API as WHATWG fetch.